chore(tests):updated tests and dependencies
