+++ requirements_dev.txt
@@ -1,5 +1,6 @@
 -r requirements.txt
 -e .
+fake-couchbase @ git+https://github.com/sitzz/python-fake-couchbase@30205d2fac74d3963738444874a05f88b14d6f84
 mypy>=1.13.0
 pytest>=5.0
 pytest-order>=1.3.0
